requirements.txt
flask==0.12.2
requests==2.16.5
python-jose==1.3.2
pycrypto==2.6.1